feat(security): drop FleetProjectRail, Patches-style project accordion
/security is now a true fleet view that mirrors /patches' grouped mode.
Each project is a collapsible accordion row (border + chevron + name +
finding-count badge + per-project Scan button on the header). Expanding
shows that project's full CVE Lite UI inline (toolbar, scan controls,
manage strip, SourcePluginCards, FixPlan, CveLiteFindings, commit banner,
confirm dialog).

All per-project state, fetches, handlers, and effects moved from
SecurityHubInner into a self-contained ProjectSecurityAccordion
component. Parent page becomes a ~95-line orchestrator: fleet-aggregate
SecurityHeader + SecuritySummaryBar (sums across projects) + a list of
accordions. FleetProjectRail no longer imported by /security (the file
itself stays in src/components/security/ for other potential consumers).

URL ?project=<id> still works: that project starts expanded on mount.
